Extract sftp and scp traits
